feat: add send_video_file and send_document_file MCP tools
Also fix pre-existing long->int conversion bugs and apply formatting fixes

+++ TelegramSearchBot/Interface/Tools/ISendDocumentToolService.cs
@@ -8,6 +8,6 @@ namespace TelegramSearchBot.Interface.Tools {
             string filePath,
             ToolContext toolContext,
             string caption = null,
-            int? replyToMessageId = null);
+            long? replyToMessageId = null);
     }
 }
